requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==4.0.3
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
ctransformers==0.2.27
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.3.0
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.9.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.3
idna==3.10
Jinja2==3.1.6
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.30
langchain-core==0.3.76
langchain-text-splitters==0.3.11
langsmith==0.4.31
llvmlite==0.43.0
MarkupSafe==3.0.3
marshmallow==3.26.1
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
openai==2.0.0
openai-whisper==20250625
orjson==3.11.3
packaging==25.0
pillow==11.0.0
propcache==0.3.2
py-cpuinfo==9.0.0
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydub==0.25.1
python-dotenv==1.1.1
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.6.2
semantic-version==2.10.0
setuptools-rust==1.12.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sympy==1.13.1
tenacity==9.1.2
tiktoken==0.11.0
tokenizers==0.22.1
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.56.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.20.1
zstandard==0.25.0
